Create common component for checkable inputs
